requirements.txt
SQLAlchemy
psycopg2
Appium-Python-Client
selenium